clovar/requirements.txt
Flask==1.0.2

# For Clover
clova-cek-sdk